requirements.txt
python-nmap
scapy
google-generativeai
streamlit
firebase-admin
python-dotenv
pandas
requests